Refactor & add mobile-menu.js module
