+++ src/index.html
@@ -14,16 +14,16 @@
       <section class="section-1 row">
         <!--Navbar for mobile, tablet & desktop screens-->
         <nav class="nav col-12 col-sm-12 col-md-12 col-lg-12 col-xl-12 col-xxl-12" id="nav">
-          <a href="#" class="d-flex d-sm-flex d-md-flex flex-row flex-sm-row flex-md-row align-items-center align-items-sm-center align-items-md-center text-decoration-none">
+          <a href="#home" class="d-flex d-sm-flex d-md-flex flex-row flex-sm-row flex-md-row align-items-center align-items-sm-center align-items-md-center text-decoration-none">
             <div class="ministry-logo-nav"></div>
             <p class="ministry-name-nav anton-font text-white mt-2">HOPE IN CHRIST MINISTRY</p>
           </a>
 
           <!--To show only when hamburger icon is clicked-->
           <ul class="nav-menu-mobile list-unstyled d-block d-sm-block d-md-block d-lg-none">
-            <li class="nav-item-mobile"><a href="#home-mobile" class="nav-link-mobile text-decoration-none familjen-grotesk-font">Home</a></li>
-            <li class="nav-item-mobile mt-3"><a href="#booking-mobile" class="nav-link-mobile text-decoration-none familjen-grotesk-font">Booking</a></li>
-            <li class="nav-item-mobile mt-3"><a href="#aboutus-mobile" class="nav-link-mobile text-decoration-none familjen-grotesk-font border-black mt-3 p-3 aboutus-nav-mobile">About Us</a></li>
+            <li class="nav-item-mobile"><a href="#home" class="nav-link-mobile text-decoration-none familjen-grotesk-font">Home</a></li>
+            <li class="nav-item-mobile mt-3"><a href="#booking" class="nav-link-mobile text-decoration-none familjen-grotesk-font">Booking</a></li>
+            <li class="nav-item-mobile mt-3"><a href="#aboutus" class="nav-link-mobile text-decoration-none familjen-grotesk-font border-black mt-3 p-3 aboutus-nav-mobile">About Us</a></li>
             <!-- <li class="nav-item-mobile"><a href="#shop-mobile" class="nav-link-mobile text-decoration-none familjen-grotesk-font fs-1 border-black mt-3 p-3">Shop</a></li> -->
             <!-- <li class="nav-item-mobile"><a href="#contactus-mobile" class="nav-link-mobile text-decoration-none familjen-grotesk-font fs-1 border-black mt-3 p-3">Contact Us</a></li> -->
           </ul>
@@ -36,9 +36,9 @@
 
           <!--To show only in laptop & larger screens-->
           <ul class="list-unstyled d-none d-sm-none d-md-none d-lg-flex flex-lg-row justify-content-lg-between">
-            <li class="nav-item-desktop mt-4 me-3"><a href="#home-desktop" class="nav-link-desktop text-decoration-none px-4 py-3 familjen-grotesk-font font-white fs-5 white-border">Home</a></li>
-            <li class="nav-item-desktop mt-4 me-3"><a href="#booking-desktop" class="nav-link-desktop text-decoration-none px-4 py-3 familjen-grotesk-font font-white fs-5 white-border">Booking</a></li>
-            <li class="nav-item-desktop mt-4 me-3"><a href="#aboutus-desktop" class="nav-link-desktop text-decoration-none px-4 py-3 familjen-grotesk-font font-white fs-5 white-border">About</a></li>
+            <li class="nav-item-desktop mt-4 me-3"><a href="#home" class="nav-link-desktop text-decoration-none px-4 py-3 familjen-grotesk-font font-white fs-5 white-border">Home</a></li>
+            <li class="nav-item-desktop mt-4 me-3"><a href="#booking" class="nav-link-desktop text-decoration-none px-4 py-3 familjen-grotesk-font font-white fs-5 white-border">Booking</a></li>
+            <li class="nav-item-desktop mt-4 me-3"><a href="#aboutus" class="nav-link-desktop text-decoration-none px-4 py-3 familjen-grotesk-font font-white fs-5 white-border">About</a></li>
           </ul>
         </nav>
 
@@ -50,7 +50,6 @@
           <p class="text-center familjen-grotesk-font text-white">Are there patterns of stagnation, addiction, relationship breakdown or financial difficulties in your family?</p>
           <p class="text-center familjen-grotesk-font text-white">Jesus Christ can set you free!</p>
         </div>
-
         
         <div class="col-4 col-sm-4 col-md-4 col-lg-4"></div>
         <a href="#booking-desktop" class="booking-btn col-4 col-sm-4 col-md-4 col-lg-4 text-decoration-none text-center familjen-grotesk-font font-black bg-yellow mx-5 px-3 py-4">BOOK A SESSION</a>
@@ -93,7 +92,7 @@
     </main>
 
     <footer class="bg-yellow d-flex d-sm-flex d-md-flex d-lg-flex flex-row flex-sm-row flex-md-row flex-lg-row justify-content-around  justify-content-sm-around justify-content-md-around justify-content-lg-around align-items-center">
-      <a href="#" class="d-flex d-sm-flex d-md-flex flex-row flex-sm-row flex-md-row align-items-center align-items-sm-center align-items-md-center text-decoration-none">
+      <a href="#home" class="d-flex d-sm-flex d-md-flex flex-row flex-sm-row flex-md-row align-items-center align-items-sm-center align-items-md-center text-decoration-none">
         <div class="ministry-logo-footer rounded-circle"></div>
         <p class="ministry-name-footer anton-font text-dark mt-2">HOPE IN CHRIST MINISTRY</p>
       </a>
